+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## vX.X.X - XXXX-XX-XX
+
+Fix leaking error on serializing `lat_max`, `lon_min` and `lon_max`
+
 ## v0.1.2 - 2024-07-27
 
 Fix error handling on not supported platform