+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## vX.X.X - XXXX-XX-XX
+## v0.1.3 - 2024-07-29
 
 - Fix leaking error on serializing `lat_max`, `lon_min` and `lon_max`
 - Fix to relax `dump` and `dumps` args type